refactor: Break UI components into seperate files
